fix: move the component to the dedicated folder
